% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{as_igraph}
\alias{as_igraph}
\title{Convert lineage to an igraph object}
\usage{
as_igraph(x, ...)
}
\arguments{
\item{x}{A \code{td_lineage} object.}

\item{...}{Unused.}
}
\value{
An \code{igraph} object.
}
\description{
Requires the \pkg{igraph} package.
}
