% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{as_json}
\alias{as_json}
\title{Machine-readable report data}
\usage{
as_json(x, pretty = TRUE, ...)
}
\arguments{
\item{x}{A \code{tdiff}, \code{td_impact} or \code{td_report} object.}

\item{pretty}{Pretty-print JSON.}

\item{...}{Unused.}
}
\value{
For \code{as_json()}, a JSON string. For \code{as_report_list()}, a list.
}
\description{
Machine-readable report data
}
