% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{td_default_rules}
\alias{td_default_rules}
\title{Default rule set}
\usage{
td_default_rules()
}
\value{
A list of \code{\link[=td_rule]{td_rule()}} objects ordered by priority.
}
\description{
Default rule set
}
