% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialdiff-package.R
\docType{package}
\name{trialdiff-package}
\alias{trialdiff}
\alias{trialdiff-package}
\title{trialdiff: Clinical Trial Data-Cut Change Detection and Impact Assessment}
\description{
A transparent, rule-based framework for detecting changes between successive data cuts of clinical trial datasets, classifying those changes into clinically meaningful categories, tracing user-defined data lineage, and assessing which downstream analyses and outputs may be affected. The package is designed to complement existing low-level data frame comparison tools by adding clinical-trial-specific classification, lineage and impact-assessment layers on top of deterministic comparison.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Hirujan-R/trialdiff}
  \item \url{https://Hirujan-R.github.io/trialdiff/}
  \item Report bugs at \url{https://github.com/Hirujan-R/trialdiff/issues}
}

}
\author{
\strong{Maintainer}: Hirujan Rangaraj \email{hirujanrangaraj@gmail.com}

Authors:
\itemize{
  \item Hirujan Rangaraj \email{hirujanrangaraj@gmail.com}
}

}
\keyword{internal}
