% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trialdiff-utils}
\alias{trialdiff-utils}
\title{Internal helpers}
\arguments{
\item{x}{An object.}

\item{nm}{A character scalar.}
}
\description{
Small, dependency-light utilities used across \code{trialdiff}.
}
\keyword{internal}
